Move strings file to source root
