+++ src/strings.ts
